+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+
+### Fixed
+- Small fixes and extra logs to saml federation parsing
+
 ## [22.0.11-1.8] - 2024-09-11
 
 ### Added